Change file names and language keys to lowercase
