+++ tests/languages/es-es.js
@@ -1,11 +1,11 @@
 var numeral = require('../../numeral'),
-    language = require('../../languages/ru-UA');
+    language = require('../../languages/es-es');
 
-numeral.language('ru-UA', language);
+numeral.language('es-es', language);
 
-exports['language:ru-UA'] = {
+exports['language:es-es'] = {
     setUp: function (callback) {
-        numeral.language('ru-UA');
+        numeral.language('es-es');
         callback();
     },
 
@@ -18,21 +18,21 @@ exports['language:ru-UA'] = {
         test.expect(16);
 
         var tests = [
-            [10000,'0,0.0000','10 000,0000'],
-            [10000.23,'0,0','10 000'],
-            [-10000,'0,0.0','-10 000,0'],
+            [10000,'0,0.0000','10.000,0000'],
+            [10000.23,'0,0','10.000'],
+            [-10000,'0,0.0','-10.000,0'],
             [10000.1234,'0.000','10000,123'],
-            [-10000,'(0,0.0000)','(10 000,0000)'],
+            [-10000,'(0,0.0000)','(10.000,0000)'],
             [-0.23,'.00','-,23'],
             [-0.23,'(.00)','(,23)'],
             [0.23,'0.00000','0,23000'],
-            [1230974,'0.0a','1,2млн'],
-            [1460,'0a','1тыс.'],
-            [-104000,'0a','-104тыс.'],
-            [1,'0o','1.'],
-            [52,'0o','52.'],
-            [23,'0o','23.'],
-            [100,'0o','100.'],
+            [1230974,'0.0a','1,2m'],
+            [1460,'0a','1k'],
+            [-104000,'0a','-104k'],
+            [1,'0o','1º'],
+            [52,'0o','52º'],
+            [23,'0o','23º'],
+            [100,'0o','100º'],
             [1,'0[.]0','1']
         ];
 
@@ -47,10 +47,10 @@ exports['language:ru-UA'] = {
         test.expect(4);
 
         var tests = [
-            [1000.234,'0,0.00$','1 000,23₴'],
-            [-1000.234,'(0,0$)','(1 000₴)'],
-            [-1000.234,'0.00$','-1000,23₴'],
-            [1230974,'(0.00a$)','1,23млн₴']
+            [1000.234,'$0,0.00','€1.000,23'],
+            [-1000.234,'($0,0)','(€1.000)'],
+            [-1000.234,'$0.00','-€1000,23'],
+            [1230974,'($0.00a)','€1,23m']
         ];
 
         for (var i = 0; i < tests.length; i++) {
@@ -78,17 +78,16 @@ exports['language:ru-UA'] = {
     },
 
     unformat: function (test) {
-        test.expect(10);
+        test.expect(9);
 
         var tests = [
-            ['10 000,123',10000.123],
+            ['10.000,123',10000.123],
             ['(0,12345)',-0.12345],
-            ['(1,23млн₴)',-1230000],
-            ['1,23млн₴',1230000],
-            ['10тыс.',10000],
-            ['-10тыс.',-10000],
-            ['23.',23],
-            ['10 000,00₴',10000],
+            ['($1,23m)',-1230000],
+            ['10k',10000],
+            ['-10k',-10000],
+            ['23er',23],
+            ['$10.000,00',10000],
             ['-76%',-0.76],
             ['2:23:57',8637]
         ];
