+++ tests/languages/fr-ca.js
@@ -1,11 +1,11 @@
 var numeral = require('../../numeral'),
-    language = require('../../languages/uk-UA');
+    language = require('../../languages/fr-ca');
 
-numeral.language('uk-UA', language);
+numeral.language('fr-ca', language);
 
-exports['language:uk-UA'] = {
+exports['language:fr'] = {
     setUp: function (callback) {
-        numeral.language('uk-UA');
+        numeral.language('fr-ca');
         callback();
     },
 
@@ -26,13 +26,13 @@ exports['language:uk-UA'] = {
             [-0.23,'.00','-,23'],
             [-0.23,'(.00)','(,23)'],
             [0.23,'0.00000','0,23000'],
-            [1230974,'0.0a','1,2млн'],
-            [1460,'0a','1тис.'],
-            [-104000,'0a','-104тис.'],
-            [1,'0o','1'],
-            [52,'0o','52'],
-            [23,'0o','23'],
-            [100,'0o','100'],
+            [1230974,'0.0 a','1,2 M'],
+            [1460,'0 a','1 k'],
+            [-104000,'0 a','-104 k'],
+            [1,'0o','1er'],
+            [52,'0o','52e'],
+            [23,'0o','23e'],
+            [100,'0o','100e'],
             [1,'0[.]0','1']
         ];
 
@@ -47,10 +47,10 @@ exports['language:uk-UA'] = {
         test.expect(4);
 
         var tests = [
-            [1000.234,'0,0.00$','1 000,23₴'],
-            [-1000.234,'(0,0$)','(1 000₴)'],
-            [-1000.234,'0.00$','-1000,23₴'],
-            [1230974,'(0.00a$)','1,23млн₴']
+            [1000.234,'0,0.00 $','1 000,23 $'],
+            [-1000.234,'(0,0 $)','(1 000 $)'],
+            [-1000.234,'0.00 $','-1000,23 $'],
+            [1230974,'(0.00 a$)','1,23 M$']
         ];
 
         for (var i = 0; i < tests.length; i++) {
@@ -64,10 +64,10 @@ exports['language:uk-UA'] = {
         test.expect(4);
 
         var tests = [
-            [1,'0%','100%'],
-            [0.974878234,'0.000%','97,488%'],
-            [-0.43,'0%','-43%'],
-            [0.43,'(0.000%)','43,000%']
+            [1,'0 %','100 %'],
+            [0.974878234,'0.000 %','97,488 %'],
+            [-0.43,'0 %','-43 %'],
+            [0.43,'(0.000 %)','43,000 %']
         ];
 
         for (var i = 0; i < tests.length; i++) {
@@ -78,18 +78,17 @@ exports['language:uk-UA'] = {
     },
 
     unformat: function (test) {
-        test.expect(10);
+        test.expect(9);
 
         var tests = [
             ['10 000,123',10000.123],
             ['(0,12345)',-0.12345],
-            ['(1,23млн₴)',-1230000],
-            ['1,23млн₴',1230000],
-            ['10тис.',10000],
-            ['-10тис.',-10000],
-            ['23.',23],
-            ['10 000,00₴',10000],
-            ['-76%',-0.76],
+            ['(1,23 M$)',-1230000],
+            ['10 k',10000],
+            ['-10 k',-10000],
+            ['23e',23],
+            ['10 000,00 $',10000],
+            ['-76 %',-0.76],
             ['2:23:57',8637]
         ];
