+++ tests/languages/uk-ua.js
@@ -1,11 +1,11 @@
 var numeral = require('../../numeral'),
-    language = require('../../languages/fr-CA');
+    language = require('../../languages/uk-ua');
 
-numeral.language('fr-CA', language);
+numeral.language('uk-ua', language);
 
-exports['language:fr'] = {
+exports['language:uk-ua'] = {
     setUp: function (callback) {
-        numeral.language('fr-CA');
+        numeral.language('uk-ua');
         callback();
     },
 
@@ -26,13 +26,13 @@ exports['language:fr'] = {
             [-0.23,'.00','-,23'],
             [-0.23,'(.00)','(,23)'],
             [0.23,'0.00000','0,23000'],
-            [1230974,'0.0 a','1,2 M'],
-            [1460,'0 a','1 k'],
-            [-104000,'0 a','-104 k'],
-            [1,'0o','1er'],
-            [52,'0o','52e'],
-            [23,'0o','23e'],
-            [100,'0o','100e'],
+            [1230974,'0.0a','1,2млн'],
+            [1460,'0a','1тис.'],
+            [-104000,'0a','-104тис.'],
+            [1,'0o','1'],
+            [52,'0o','52'],
+            [23,'0o','23'],
+            [100,'0o','100'],
             [1,'0[.]0','1']
         ];
 
@@ -47,10 +47,10 @@ exports['language:fr'] = {
         test.expect(4);
 
         var tests = [
-            [1000.234,'0,0.00 $','1 000,23 $'],
-            [-1000.234,'(0,0 $)','(1 000 $)'],
-            [-1000.234,'0.00 $','-1000,23 $'],
-            [1230974,'(0.00 a$)','1,23 M$']
+            [1000.234,'0,0.00$','1 000,23₴'],
+            [-1000.234,'(0,0$)','(1 000₴)'],
+            [-1000.234,'0.00$','-1000,23₴'],
+            [1230974,'(0.00a$)','1,23млн₴']
         ];
 
         for (var i = 0; i < tests.length; i++) {
@@ -64,10 +64,10 @@ exports['language:fr'] = {
         test.expect(4);
 
         var tests = [
-            [1,'0 %','100 %'],
-            [0.974878234,'0.000 %','97,488 %'],
-            [-0.43,'0 %','-43 %'],
-            [0.43,'(0.000 %)','43,000 %']
+            [1,'0%','100%'],
+            [0.974878234,'0.000%','97,488%'],
+            [-0.43,'0%','-43%'],
+            [0.43,'(0.000%)','43,000%']
         ];
 
         for (var i = 0; i < tests.length; i++) {
@@ -78,17 +78,18 @@ exports['language:fr'] = {
     },
 
     unformat: function (test) {
-        test.expect(9);
+        test.expect(10);
 
         var tests = [
             ['10 000,123',10000.123],
             ['(0,12345)',-0.12345],
-            ['(1,23 M$)',-1230000],
-            ['10 k',10000],
-            ['-10 k',-10000],
-            ['23e',23],
-            ['10 000,00 $',10000],
-            ['-76 %',-0.76],
+            ['(1,23млн₴)',-1230000],
+            ['1,23млн₴',1230000],
+            ['10тис.',10000],
+            ['-10тис.',-10000],
+            ['23.',23],
+            ['10 000,00₴',10000],
+            ['-76%',-0.76],
             ['2:23:57',8637]
         ];
